fix(components): move enum to another file
